requirements.txt
# Core
pandas>=2.2,<3.0
numpy>=1.26,<3.0
PyYAML>=6.0.1,<7.0
langdetect>=1.0.9,<2.0
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0

# Test ---
pytest>=8.0,<9.0

# EDA 
matplotlib>=3.8,<4.0
seaborn>=0.13,<1.0


tqdm>=4.66,<5.0            
notebook>=7.2,<8.0         


sentence-transformers>=3.0,<4.0
transformers>=4.44,<5.0

 
scikit-learn>=1.5,<2.0



rouge-score>=0.1.2
bert-score>=0.3.13         

#  gRPC + Event
grpcio>=1.64,<2.0
grpcio-tools>=1.64,<2.0
protobuf>=6.31,<7.0
pika>=1.3,<2.0             
pika>=1.3,<2.0
psycopg2-binary>=2.9,<3.0   
